pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
ext-modules = [
  {name = "jpamb.timer", sources = ["jpamb/timer.c"]}
]

[tool.setuptools.packages.find]
include = ["jpamb", "jpamb.jvm"]  


[project]
name = "jpamb"
version = "0.3.0"
dependencies = [
  "PyYAML==6.0.*",
  "click==8.1.*",
  "hypothesis>=6.136",
  "loguru==0.7.*",
  "matplotlib==3.10.*",
  "tree-sitter-java >= 0.23, <= 0.26",
  "tree-sitter >= 0.23, < 0.26",
  "z3-solver==4.15.*",
]
authors = [
  { name = "Christian Gram Kalhauge", email = "chrg@dtu.dk" }
]
description = "Utilities for working with the jpamb benchmark suite"
requires-python = ">=3.13"
license = "BSD-3-Clause"
classifiers = [
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
stats = [
  "numpy==2.1.1",
  "pandas==2.2.2",
  "plotly==5.24.0",
]

[project.scripts]
jpamb = "jpamb.cli:cli"

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]

[tool.pytest.ini_options]
markers = [
  "slow: mark test as slow"
]

